# ENH: Speed-up MovieStim (noticeable for larger videos)