See https://github.com/psychopy/psychopy/pull/6439 for context. With my AMD+Ubuntu 22.04 laptop and [this sample video](https://github.com/psychopy/psychopy/pull/6439#issuecomment-2100781642), I get an overall reduction in `MovieStim.draw()` time (starting from [this patch](https://github.com/psychopy/psychopy/pull/6448)) from ~34ms to ~4ms (i.e. ~40ms -> 4ms across both sets of patches). I haven't tested on other systems yet, but will try on a Windows 11 + Intel and Ubuntu 20.04 + AMD desktop today.